remove ListenerId utility class; move functionality to KEXPPower directly
